Initial extension source commit
manifest.json defines which files should be loaded.
background.js is the entry point of the extension.
popup.html is the GUI of the extension.

For now the extension only initializes the different modules, creates
the database if it doesn't exist, allow to clear the database, and
monitor the navigation events occuring (log in the javascript console
all the urls of the pages that have been visited).

message passing design:
Messages between the GUI(popup.js) and the background extension are
handled by background.js which redirects the messages according to their
specified module.