build a JSON representation of a webpage
The background script of extensions can't access the DOM of the web
pages.

A script is injected in every web pages loaded in the browser and
recurcively builds a JSON representation of the page. Each element/node
is represented as:

{tag:"tag of the node",
attributes: {attribute1: "value 1", attribute2:"value2"},
children: [{child1},{child2}],
text: "text content of the element"
}

Empty attributes are discared.

When a page is loaded, the navigator listener send a message to the
injected script which JSONifies each element and return the JSON object
to the background.js, which himself send the response to the callback
function specified earlier.