bug fix benchmark
the primary key used in the bechmark was a timestamp. Since the objects
were created programmatically, it occured that two of them were created
with the same timestamp. the second would overwrite the first one hence
the lower number of entry in the store. in the frame of the benchmark,
the primary key is an autoincrementing value