FIX: storage manager query on primary key possible
The request can now be made on the primary key rather than secondary
indexes. For that the storage manager first test if the key is the name
of an index. If not it MUST be the primary key (otherwise an error will
be thrown).

Implementation of a count function to know the number en entries in a
objectstore