Store each page visited in temporary store
Flow: Each page that a user visit receive the injectable script which
create a JSON version of the DOM, and send it to background.js which
redirect it to the appropriate class: DocumentPreprocessorPool. This
pool create an instance of a DocumentPreprocessor which will take care
of retreiving features and storing the object in the temporary store.
For now, only the timestamp, the URL of the document, the language of
the page, the DOM and a screenshot are stored into the store.