stores user events
The extension now listen to some javascript events such as the browser
clicking the page or browsing away and stores these events in the
databases in the related record. The Preprocessor pool is extended for
each new page. If an "unload" event is triggered, the preprocessor pool
will delete the preprocessor responsible for the page which triggered
this event.