capture highlighted text
the injectable script listen to mouseup event (last part of a click) and
check if the windo.getSelection() changed. If it did, then the user
selected a new piece of text, this text is stored in a user event called
highlight. This info should be   of some importance when it come to
analyze what the user focused on.