Info gatherer for the user study
One of the requirements  from the user study is the possibility to get a
picture of the user and his location at the beginning of each session.

This commit is a first step toward this goal. A pages can be opened from
the extension which allow the user to:
- take a picture of him
- locate himself on google map
- label his location with an already defined label or create one if none
fits

When validating, the page send these information along with a timestamp
to the userStudyToolbox component which stores them into the
userStudySession object store.

If the user doesn't want her session to be recorded, she can specifies
so, then the information is not gathered and the extension is set to
incognito mode.