fix: improve database availability management
The database takes a few milliseconds to initialize. In the meanwhile,
other components are initializing as well and some of them are making
database queries. Now, query are buffered in a queue (javascript array)
which is emptied as soon as the database is ready. At the same time,
this commit enforce the sequenciality of the query at the software level
(which is not mandatory since indexedDB is suppose to handle it
anyway)).