pinned session information page
The page to gather information about user session (picture and location)
is now pinned to make it harder for the user to ignore it. Several
additional actions can be taken to ensure that the user fill the form
in: forbid browsing away from this page, forbid switching tab. These
actions will be implemented if needed.