refactoring: parameters database
the domain specific rules has been moved in a "parameters" database.
This database will be used in later commits to include additional
parameters for the extension. (technical detail: rules are not stored as
array anymore, but as object's attribut. To access them a <for key,value
of object> must be used to iterate through them)