storage of the url of website used in user study
The lsit of the website selected for the user study must be stored in
indexedDB. The lsit of the websites can be modified from the option page
and is stored in the "parameters" database.

The URL can be represented as a regular expression to allow to fetch
pages from subdomains etc.