Foreign key impl. + index naming convention
A system of foreign key has been implemented in indexedDB in order to
link objects between them in a relational database fashion.

For that, the child object must possess an index pointing toward the
primary key of the parent. (by convention the name of the FK is _PK. For
instance the object child will have an attribute _parentId = "12345")

To retrieve a parent object with its children, one has to append the
method .getChildren() at the end of a query.

this method takes an array as parameter containing a liste of the
children that we want to fetch along with the parent object. the array
is defined as follow: [{name:"child1",
objectStore:"children1"},{name:"child2", objectStore:"children2"}]

the parent object will be returned with the following structure:
{
field1:1,
field2:2,
child1:{child1},
child2:{child2},
_children:[{name:"child1", objectStore:"children1"},{name:"child2",
objectStore:"children2"}]
}

the retrieval and the storage of the children are done in a recursive
fashion (the deletion is not yet supported and would leave orphans in
the database)