feature: dump for user study
The userStudyToolbox contains a dump method which fetch the list of the
websites (prefixes) that are part of the user study and fetch the
relative objects in the database along with their children and the
session they belong to. A (quite huge) JSON is created and is then sent
via POST XHR to a server.

The next problem to tackle would be to split this JSON into several of
standard http size (a few KB), in order to not preserve the RAM of the
user.