screenshot resistance
This feature might be cancelled in a later commit due to its "none user
friendly  experience". Indeed, this feature ensure that the screenshot
taken is effectively the one of the page that just loaded, on the other
hand, it requires that the loaded page in activated in the focused
windows. which leads to a blinking window when the user switches tabs
before a page is loaded. (ie, the extension activate the loaded page
tab, take the screenshot, and then goes back to the tab the user was
on.)