End of bootstrap migration (context page)
This commit represent the end of the migration toward the twitter
bootstrap styling.

The context page has been adapted to the bootstrap css. From now, only
minor modifications should be required regarding the styling.

NB: the bootstrap.css has been appended with global classes definition
such as centered, bold, etc. as well as two classes needed for the
layout of the two column items (picture/validation pages) which could
not be obtain with the span6 classes already in place.

This extra classes has been added at the end of the css file in a custom
section.