refactoring: processors are now "per tabs"
Processor are no per tabs. each time a page is loaded, the
processorspool detect the tab in which it is open. If no processor has
been created for this tab or if the processor exist but is processing a
different page (different URL), a new one is created and provided with
the info of the new page. Otherwise, the processor is just updated with
the new page DOM and return the pageId of the previous instance of this
page (only way to pass the pageId between refresh of a page).

This refactoring allows to handle the refresh of a page (provided the
URL stays stricly the same).

(( This commit should also ease the refactoring of the screenshot method
which will simply detect at the creation of the documentPreprocessor if
the page is on the activated page. If yes the screenshot is taken, if
not it waits, and the next time the tab of the page is activated, the
documentPreprocessor is notified and the screenshot is taken. (this is
based on the idea that there is no need to capture the page that the
user opened but didn't even see). ))