FIX: screenshot is now taken without flickering
The former method to take screenshot of webpages was implying a
flickering of the windows. To avoid this issue, the screenshot is now
taken as soon as the page is activated. If the tab was switched from
when the page was loaded, the extenstion will wait for the page to be
refocused to take the screen (anyway there is now meaning in taking the
screenshot of a page that the user didn't even look at).