fix:storage of objet containing children
The storage of element with children was not done properly: I didn't
take into account the case where children were composed of several
elements. ie. a page has a child "userActions" which can contain more
than one event.

To go around, when a children is fetched, it is automatically stored in
an array (the array might be of size 1). And now, when storing the
object back, we iterate over every children and on each of them we
iterate over the different elements of the child. typical structur:
parent{child1{el1, el2},child2{el1, el2, el3}, child3{el1}}