population of the inspector
Populate the template using the angularjs templating system. This might
not be the right way but for now it works: The controller create a
future object and start a database query via message passing to the
userStudyToolbox. When receiving the result, the future object is
resolved(updated). All of that happening inside a scope.$apply method in
order fo angularjs to notice that there has been an update.