Daily and smart (by session) dump
Dump are now taking daily (or asap after 24 hours). The dump is composed
of user sessions which contains pages. The dump is then sent to a server
via a POST request. The address of the server is (for now) hardcoded at
the head of userStudyToolbox.coffee. Modification have been done also to
the regular dump to match the same json structure. The loading of
mem0r1es has been updating accordingly.