The referer of each page as an mem0r1es attribute
The referer of each page is catched from the http request sent by the
browser, it is temporarily stored in an documentPreprocessorpool's
object (@referersList), until the page is actually fetched and rendered.
When the document preprocessor is created, it is provided with the
matching referer.

A listcleaner is called every 5 minutes to make sure that referers that
have not been claimed don't poluate the memory (to prevent memory
leakage).