capture of idle and focus time series
Using the chrome idle API, we now store in each mem0r1e a list of its
states regarding the activity of the user : if she has been idle for
more than 5 min, as well as the focus of the page: if the page has lost
focus while it was open. The 5 minutes idle parameter must be tweaked in
both documentProcessor and documentProcessorPool.