fix: prevent several session pages
Now we check if a session page is already up (in the frame of the
automatic session update, not the manual updates). If yes then we don't
display the page, if not then we display it. Like that even after hours
of idle time, there should be at most one session tab.