The database versionUpdate is now correct
So far, chrome was not supporting the onupgradeneeded event when opening
a database. In the new version of chrome (25) it seems to cause some
problem regarding the loading of the database.

We solve this issue by using the onupgradeneeded  event which is the
normalized way to update an indexedDB database.