load database dump via local file system
As in previous commit, now we copy first the JSON to the chrome file
system before processing it. NB: for large file their is still crashes
probably due to the number of time the JSON is represented in memory in
userStudyToolbox while procesing the file: 1 arrayBuffer to load the
file + 1 int8 array + 1 regular array +1 string. We should avoid crashes
by chunking the reading phase ( loop over the array buffers to read into
the file).